+++ CHANGES.rst
@@ -24,6 +24,8 @@ Changelog
 
 **Changed**
 
+- #1215 Do not copy CaptureDate and Result in retest analyses when created
+- #1215 Do not modify the ID of analysis on retraction
 - #1207 Make use of adapters for instrument auto-import
 - #1206 Make use of adapters for instrument import/export interfaces
 - #1203 Remove explicit definition of transitions in AR listing
@@ -91,6 +93,8 @@ Changelog
 **Fixed**
 
 - #1214 Disallow entry of analysis results if the sample is not yet received
+- #1213 Fix instrument notification display in Manage Results View
+- #1212 Fix typo in SamplingFieldsVisibility
 - #1191 Some worksheets pre-1.3 with published analyses remain in open/to_be_verified state
 - #1190 Fixed evolution chart for reference analyses
 - #1183 Fix results calculation of dependent calculations